change HTML + refresh gitignore

+++ Wolf Engine 2.0/Mesh.h
@@ -32,7 +32,7 @@ namespace Wolf
 		m_indexBuffer.reset(new Buffer(sizeof(uint32_t) * indices.size(), VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, UpdateRate::NEVER));
 		m_indexBuffer->transferCPUMemoryWithStagingBuffer((void*)indices.data(), sizeof(uint32_t) * indices.size());
 
-		m_indexCount = indices.size();
+		m_indexCount = static_cast<uint32_t>(indices.size());
 	}
 }
